chore: api extension 설정
