+++ src/pages/api/auth/[...nextauth].api.ts
